[refactor] move omniauth login into service object
