+++ spec/services/games_creator_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe Games::Creator do
 
-  let(:lobby)   { create(:lobby) }
+  let(:lobby)   { create :lobby }
   let(:creator) { Games::Creator.new(lobby) }
 
   describe "#create_new_game" do
