+++ spec/services/score_updater_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe Lobbies::ScoreUpdater do
 
-  let(:user) { create :user }
+  let(:user)  { Users::Creator.new.create_new_user(attributes_for :user) }
   let(:lobby) { Lobbies::Creator.new.create_new_lobby(user) }
   let(:score_updater) { Lobbies::ScoreUpdater.new(lobby) }
