+++ spec/services/sets_creator_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe Sets::Creator do
 
-  let(:user)    { create :user }
+  let(:user)    { Users::Creator.new.create_new_user(attributes_for :user) }
   let(:lobby)   { create :lobby }
   let(:game)    { Games::Creator.new(lobby).create_new_game }
   let(:creator) { Sets::Creator.new(game) }
